--- src/main/java/ByeCommand.java
@@ -1,7 +1,0 @@
-public class ByeCommand extends Command {
-    @Override
-    public void execute(TaskList tasks,Ui ui,Storage storage) {
-        ui.showBye();
-        this.setExit();
-    }
-}
