--- src/main/java/Command.java
@@ -1,13 +1,0 @@
-public abstract class Command {
-    boolean isExit = false;
-
-    public boolean isExit() {
-        return this.isExit;
-    }
-
-    public void setExit() {
-        this.isExit = true;
-    }
-
-    public abstract void execute(TaskList tasks,Ui ui,Storage storage) throws DukeException;
-}
