+++ src/main/java/duke/command/AddCommand.java
@@ -1,3 +1,12 @@
+
+package duke.command;
+
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.task.Task;
+
 public class AddCommand extends Command {
     private Task task;
 
@@ -6,7 +15,7 @@ public class AddCommand extends Command {
     }
 
     @Override
-    public  void execute(TaskList tasks,Ui ui,Storage storage) {
+    public  void execute(TaskList tasks, Ui ui, Storage storage) {
         tasks.add(task);
         int total = tasks.totalSize();
         ui.showAdd(task,total);
