+++ src/main/java/duke/command/DefaultCommand.java
@@ -1,0 +1,13 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class DefaultCommand extends Command {
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+        throw new DukeException("OOPS!!! I'm sorry, but I don't know what that means :-(\n");
+    }
+}
