+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.task.Task;
+
 public class DeleteCommand extends Command {
     private int number;
 
@@ -5,7 +12,7 @@ public class DeleteCommand extends Command {
         this.number = number;
     }
     @Override
-    public void execute(TaskList tasks,Ui ui,Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
         Task task = tasks.getTaskList().get(this.number - 1);
         tasks.delete(this.number - 1);
         int total = tasks.totalSize();
