+++ src/main/java/duke/task/Deadline.java
@@ -1,3 +1,6 @@
+package duke.task;
+
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
