+++ src/main/java/duke/task/Event.java
@@ -1,3 +1,6 @@
+package duke.task;
+
+
 public class Event extends Task {
     private String time;
     private String name;
