+++ text-ui-test/runtest.sh
@@ -13,7 +13,9 @@ then
 fi
 
 # compile the code into the bin folder, terminates if error occurred
-if ! javac -cp ../src/main/java -Xlint:none -d ../bin ../src/main/java/*.java
+if  ! javac -cp ../src/main/java/duke/command -Xlint:none -d ../bin ../src/main/java/duke/command/*.java &&
+    ! javac -cp ../src/main/java/duke/task -Xlint:none -d ../bin ../src/main/java/duke/task/*.java &&
+    ! javac -cp ../src/main/java/duke -Xlint:none -d ../bin ../src/main/java/duke/*.java
 then
     echo "********** BUILD FAILURE **********"
     exit 1
